+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CorrDiff CMIP6->ERA5 diagnostic model
 - Added derived TCWV diagnostic model
 - Added verbose option for inference workflows in run.py to toggle off tqdm
+- Added Atlas prognostic model
 - Added StormScope GOES and MRMS models
 
 ### Changed